chore(Satismeter): move sdk loader to different file
